requirements.txt
streamlit>=1.31,<1.39
pillow>=10.0.0
numpy>=1.23
opencv-python-headless>=4.9.0.80
pytesseract>=0.3.10
dateparser>=1.2.0
pandas>=2.2.0